requirements.txt
pytest==8.2.0
pytest-cov>=4.1.0
ruff>=0.1.0
mypy>=1.7.0
pre-commit>=3.5.0
loguru>=0.7.2
PyYAML==6.0.1
mutagen==1.47.0
watchdog==4.0.1
Pillow==10.4.0
pydub==0.25.1
numpy==1.26.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-barcode==0.16.1
rich>=13.0.0

# Optional dependencies (for enhanced features)
# openai>=1.0.0  # For AI-generated titles and YouTube descriptions (optional)

# YouTube API dependencies (for video upload)
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0